requirements.txt
google-genai
pandas


